Just depend on Syn with default features
We used to attempt to disable Syn's clone-impls feature because it was
not needed for proc-macro-hack, but Cargo seems to enable it anyway
because one of our dev-dependencies uses it. Instead of figuring out how
to test a non-Cargo build to avoid accidentally using clone-impls, just
depend on Syn's default feature set because downstream code is probably
doing the same anyway.